+++ CHANGELOG.md
@@ -26,6 +26,7 @@ All notable changes to this project will be documented in this file.
 - Add ListBuilder for building lists by reflection (#705)
 - F# example (#699)
 - Support for discontinuities in AreaSeries (#215)
+- Support for Windows Universal 10.0 apps (#615)
 
 ### Changed
 - Renamed OxyPlot.WindowsUniversal to OxyPlot.Windows (#242)